console: Move application selectors to new file move selectors to selectors/applications.js
