+++ pkg/webui/console/containers/application-payload-formatters/downlink.js
@@ -29,8 +29,8 @@ import { updateApplicationLinkSuccess } from '../../store/actions/link'
 import {
   selectApplicationIsLinked,
   selectApplicationLinkFormatters,
-} from '../../store/selectors/application'
-import { selectSelectedApplicationId } from '../../store/selectors/applications'
+  selectSelectedApplicationId,
+} from '../../store/selectors/applications'
 
 import api from '../../api'
