+++ pkg/webui/console/store/selectors/applications.js
@@ -12,16 +12,37 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { GET_APPS_LIST_BASE } from '../actions/applications'
-import { GET_APP_BASE } from '../actions/application'
+import {
+  GET_APPS_LIST_BASE,
+  GET_APPS_RIGHTS_LIST_BASE,
+} from '../actions/applications'
+import {
+  GET_APP_BASE,
+  GET_APP_API_KEY_BASE,
+  GET_APP_API_KEYS_LIST_BASE,
+} from '../actions/application'
+import { GET_APP_LINK_BASE } from '../actions/link'
+
 import {
   createPaginationIdsSelectorByEntity,
   createPaginationTotalCountSelectorByEntity,
 } from './pagination'
+import {
+  createEventsSelector,
+  createEventsErrorSelector,
+  createEventsStatusSelector,
+} from './events'
+import {
+  createRightsSelector,
+  createUniversalRightsSelector,
+} from './rights'
+import { createApiKeysSelector } from './api-keys'
+import { createApiKeySelector } from './api-key'
 import { createFetchingSelector } from './fetching'
 import { createErrorSelector } from './error'
 
 const ENTITY = 'applications'
+const ENTITY_SINGLE = 'application'
 
 // Application
 export const selectApplicationStore = state => state.applications
@@ -42,3 +63,47 @@ export const selectApplications = state => selectAppsIds(state).map(id => select
 export const selectApplicationsTotalCount = state => selectAppsTotalCount(state)
 export const selectApplicationsFetching = state => selectAppsFetching(state)
 export const selectApplicationsError = state => selectAppsError(state)
+
+// Events
+export const selectApplicationEvents = createEventsSelector(ENTITY)
+export const selectApplicationEventsError = createEventsErrorSelector(ENTITY)
+export const selectApplicationEventsStatus = createEventsStatusSelector(ENTITY)
+
+// Rights
+export const selectApplicationRights = createRightsSelector(ENTITY)
+export const selectApplicationUniversalRights = createUniversalRightsSelector(ENTITY)
+export const selectApplicationRightsError = createErrorSelector(GET_APPS_RIGHTS_LIST_BASE)
+export const selectApplicationRightsFetching = createFetchingSelector(GET_APPS_RIGHTS_LIST_BASE)
+
+// Api Keys
+export const selectApplicationApiKeys = createApiKeysSelector(ENTITY)
+export const selectApplicationApiKeysError = createErrorSelector(GET_APP_API_KEYS_LIST_BASE)
+export const selectApplicationApiKeysFetching = createFetchingSelector(GET_APP_API_KEYS_LIST_BASE)
+export const selectApplicationApiKey = createApiKeySelector(ENTITY_SINGLE)
+export const selectApplicationApiKeyFetching = createFetchingSelector(GET_APP_API_KEY_BASE)
+export const selectApplicationApiKeyError = createErrorSelector(GET_APP_API_KEY_BASE)
+
+// Link
+const selectLinkStore = state => state.link
+export const selectApplicationLink = state => selectLinkStore(state).link
+export const selectApplicationLinkStats = state => selectLinkStore(state).stats
+export const selectApplicationLinkFetching = createFetchingSelector(GET_APP_LINK_BASE)
+export const selectApplicationLinkError = createErrorSelector(GET_APP_LINK_BASE)
+export const selectApplicationLinkFormatters = function (state) {
+  const link = selectApplicationLink(state) || {}
+
+  return link.default_formatters
+}
+export const selectApplicationIsLinked = function (state) {
+  const linkStore = selectLinkStore(state)
+  const link = selectApplicationLink(state) || {}
+  const error = selectApplicationLinkError(state)
+  const stats = selectApplicationLinkStats(state)
+
+  const hasBase = Boolean(link.api_key)
+  const hasError = Boolean(error)
+  const isLinked = linkStore.linked
+  const hasStats = Boolean(stats)
+
+  return hasBase && !hasError && isLinked && hasStats
+}
