+++ pkg/webui/console/views/application-api-key-edit/index.js
@@ -28,8 +28,8 @@ import { ApiKeyEditForm } from '../../../components/api-key-form'
 
 import { getApplicationApiKey } from '../../store/actions/application'
 import { getApplicationsRightsList } from '../../store/actions/applications'
-import { selectSelectedApplicationId } from '../../store/selectors/applications'
 import {
+  selectSelectedApplicationId,
   selectApplicationRights,
   selectApplicationUniversalRights,
   selectApplicationRightsError,
@@ -37,7 +37,7 @@ import {
   selectApplicationApiKey,
   selectApplicationApiKeyError,
   selectApplicationApiKeyFetching,
-} from '../../store/selectors/application'
+} from '../../store/selectors/applications'
 
 import api from '../../api'
