+++ pkg/webui/console/views/application-collaborator-add/index.js
@@ -34,13 +34,13 @@ import SubmitBar from '../../../components/submit-bar'
 import RightsGroup from '../../components/rights-group'
 
 import { getApplicationsRightsList } from '../../store/actions/applications'
-import { selectSelectedApplicationId } from '../../store/selectors/applications'
 import {
+  selectSelectedApplicationId,
   selectApplicationRights,
   selectApplicationUniversalRights,
   selectApplicationRightsFetching,
   selectApplicationRightsError,
-} from '../../store/selectors/application'
+} from '../../store/selectors/applications'
 
 import api from '../../api'
