+++ pkg/webui/console/views/application-payload-formatters/index.js
@@ -33,8 +33,8 @@ import {
   selectApplicationIsLinked,
   selectApplicationLink,
   selectApplicationLinkFetching,
-} from '../../store/selectors/application'
-import { selectSelectedApplicationId } from '../../store/selectors/applications'
+  selectSelectedApplicationId,
+} from '../../store/selectors/applications'
 
 import style from './application-payload-formatters.styl'
